requirements.txt
py~=1.11.0
pip~=21.3.1
wheel~=0.37.1
retry~=0.9.2
python-nmap~=0.7.1
requests~=2.28.1
attrs~=22.1.0
pytest~=7.2.0
decorator~=5.1.1
pluggy~=1.0.0
exceptiongroup~=1.0.0
iniconfig~=1.1.1
certifi~=2022.9.24
idna~=3.4
urllib3~=1.26.12
pyparsing~=3.0.9
setuptools~=60.2.0
msgpack~=1.0.4
fontawesome~=5.10.1.post1
tqdm~=4.64.1